FE & BE - Add exception for adding and removing students

+++ src/main/java/com/alex/studentmanagementsystem/student/StudentService.java
@@ -1,6 +1,8 @@
 package com.alex.studentmanagementsystem.student;
 
+import com.alex.studentmanagementsystem.exception.ApiRequestException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import java.util.List;
 
@@ -14,10 +16,23 @@ public class StudentService {
     }
 
     public void addStudent(Student student) {
+        if (studentRepository.existsByEmail(student.getEmail())) {
+            throw new ApiRequestException(
+                    String.format("The student with email %s already exists", student.getEmail())
+            );
+        }
+
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
+        if (!studentRepository.existsById(studentId)) {
+            throw new ApiRequestException(
+                    String.format("The student with id %d doesn't exist", studentId),
+                    HttpStatus.NOT_FOUND
+            );
+        }
+
         studentRepository.deleteById(studentId);
     }
 }
